Rework the protocols for changing window size
Now the window size can be changed using a configure request in
the wl_surface, but the actual change (if any) is decided by the
compositor and communicated using a configure_notify event. Thus,
attaching buffers no longer affects the size of a surface.
Also, move and resize requests in wl_shell have been consolidated
in a single start_move_resize (equivalent to _NET_WM_MOVERESIZE
in X11) and two requests, resize and restack, have been added, for
controlling the position of own windows.