Introduce binding annotations for manual bindings
Sometimes the generated bindings are not enough to provide a
certain feature. This commit introduces a client:custom attribute
that can be applied to interface and requests and blocks all
generated code. It does not affect server side bindings.
This can be used for wl_display, and in the future for some caching
in wl_surface.