Add caching and getters/setters to client-side properties
Introduce an additional mode for wayland-scanner, that generates
client-side glue code, holding the definition of structures,
of constructors and destructors. This requires a new header file,
wayland-client-private.h, which defines wl_proxy; it is installed
for extension libraries.
wl_display_bind and wl_proxy_create are no longer exposed. Instead,
client stubs call real functions specific of each interface, that
pass the size of the structure. Inside the constructor, a listener
for property changes is installed.
All properties are exposed as a getter/setter pair, which operate
on the cached value as well as the protocol side.