Introduce a method for obtaining client credentials
This commit adds a special wl_display.hello request that is
emitted synchronously when opening the connection, and is immediately
read by the compositor. This request is accompained with a control
message of type SCM_CREDENTIALS, holding pid, uid and gid of the
client, which are stored in wl_client for later use.