Improve support for versioning
Now upon connection the client must supply the protocol and
wl_display expected version in the handshake (the wl_display.hello
request). If any doesn't match, connection is immediately terminated.
Also, when creating a new client object, a version is supplied
taken from the wl_interface, transparent to the library user), that
the compositor can check in bind or in the creation method implementation.
Finally, name 1 (the wl_display) can no longer be bound using a
manual bind request; instead, it is bound automatically to id 1
as part of the handshake.